add a function to compute traj max distance
