+++ src/trajectory.h
@@ -11,13 +11,18 @@
 int get_augmentation_number(const LaunchConfig* launch, FunctionStatus* status);
 
 /**
- * A function to get all trajectories that are feasible to hit the target.
- * @param launchConfigArr represents the launch configuration whose angle to be computed to hit the target.
- * @param lengthOfArr represents the length of the launch configuration array.
- * @param distance of the target from the mortar.
- * @return the status of the function result.
+ * A function to get the trajectory height at a distance from the mortar.
+ * @param launchConfig represents the mortar's launch configuration.
+ * @param distance represents the distance from the mortar.
+ * @return the trajectory height.
  */
-FunctionStatus get_all_trajectories_that_hit_target_by_speed(LaunchConfig* launchConfigArr, int lengthOfArr,
-                                                             double distance);
+double get_trajectory_height_by_distance(const LaunchConfig *launchConfig, double distance);
 
+/**
+ * A function to get the trajectory maximum distance.
+ * @param launchConfig represents the mortar's launch configuration.
+ * @param status represents the status of the function.
+ * @return the maximum distance of the trajectory.
+ */
+double get_trajectory_max_distance(const LaunchConfig* launchConfig, FunctionStatus* status);
 #endif
