+++ tests/test_trajectory.c
@@ -21,26 +21,23 @@ void test_get_augmentation_number() {
 }
 
 /**
- * A function to test that all trajectories that hit the target can be retrieved.
+ * A function to test the trajectory height computed from a given launch config distance from the mortar.
  */
-void test_get_all_trajectories_that_hit_target_by_speed(){
-    int index;
-    LaunchConfig launchConfigArr[12] = {create_test_launch_config(0, 125),
-                                       create_test_launch_config(0, 115),
-                                       create_test_launch_config(0, 110),
-                                       create_test_launch_config(0, 41),
-                                       create_test_launch_config(0, 30),
-                                       create_test_launch_config(0, 0)};
-    double solutionAngles[] = {2.985433, 87.014567, 3.529759, 86.470241, 3.859854, 86.140146, 37.63689, 52.36311};
-    double solutionVelocities[] = {125, 125, 115, 115, 110, 110, 41, 41};
-    double distance = 165.75;
-    FunctionStatus status = get_all_trajectories_that_hit_target_by_speed(launchConfigArr, 12, distance);
+double test_get_trajectory_height_by_distance() {
+    LaunchConfig launch = create_test_launch_config(37.5, 110);
+    double distance = 20.5;
+    double height = (distance*tan(PI/180*launch.angle)) -
+                    (0.5*GRAVITY_ACCEL*pow(distance, 2)/pow(launch.velocity*cos(PI/180*launch.angle), 2));
+    TEST_ASSERT_EQUAL_FLOAT(height, get_trajectory_height_by_distance(&launch, distance));
+}
+
+/**
+ * A function to test that the max distance of a trajectory can be computed.
+ */
+void test_get_trajectory_max_distance() {
+    LaunchConfig launch = create_test_launch_config(30.0, 115);
+    double max_distance = pow(launch.velocity, 2)/GRAVITY_ACCEL;
+    FunctionStatus status;
+    TEST_ASSERT_EQUAL_FLOAT(max_distance, get_trajectory_max_distance(&launch, &status));
     TEST_ASSERT_EQUAL(true, status.success);
-    for (index=0; index < 12; index++) {
-        if (launchConfigArr[index].velocity == 0 && fabs(launchConfigArr[index].angle) < FLOATING_POINT_PRECISION) {
-            break;
-        }
-        TEST_ASSERT_EQUAL_INT(solutionVelocities[index], launchConfigArr[index].velocity);
-        TEST_ASSERT_EQUAL_FLOAT(solutionAngles[index], launchConfigArr[index].angle);
-    }
 }
